Fixes directory listing parsing for large directories
When listing a directory with several files, the information will be available to the stream in multiple parts (by multiple calls to the stream:handlEvent method). Sometimes this information will be incomplete and won't be completely parsed.

If CFFTPCreateParsedResourceListing can't parse all the entries in the listing, the remaining bytes in the buffer must be stored to be parsed together when the stream has new bytes available.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>